Drivers! Drivers build standalone with libgcc and incbin.
- Added libgcc _start method that calls main
- All drivers are built standalone into their own executables.
- All drivers are built into src/incbin, then incbin includes them
into the kernel's memory.

+++ src/kbd.cpp
@@ -32,7 +32,7 @@ void kbd_main() {
 }
 
 int kbd_irq(kisrcall_t *info) {
-	int key = io_inb(0x60);
+	unsigned int key = (unsigned int)io_inb(0x60);
 
 	switch (key) {
 	// L shift release
